Properly compute workspace members
Do not simply try subdirectories.
Workspace members can actually be defined in the [workspace] section of Cargo.toml
and can contain globs.
We now properly compute these paths, and remove excludes if any are specified.
This is a small but important change, because previously, many workspace members
were skipped because we were not doing this.