Make it possible to exhaust all possibilities
Not just return as soon as a building fuzz target is generated.

Pass --stop-on <bug/useful/building> and --max-fuzz-targets <number> to choose when to stop
For example --stop-on bug --max-fuzz-targets 2 would stop as soon as 2 bugs are found
for this code base.

Note that the results (json) format changed and that eval_results
currently won't work.